refactor: moved checkbox and radio and fixed them
