--- src/shared/ui/Radio/index.ts
@@ -1,3 +1,0 @@
-import Radio from './Radio'
-
-export { Radio }
