+++ src/shared/ui/input/Radio/Radio.scss
